Securing the Recipe Vault Web Application
In this project, I deployed and assessed a simple web application environment’s security posture and also tested the security of the environment by simulating attack scenarios and exploiting cloud configuration vulnerabilities. Implemented monitoring to identify insecure configurations and malicious activity. Applied methods learned in the course to harden and secure the environment and designed a DevSecOps pipeline.